feat: implement ServiceFreshAddress and integrate with existing services
